requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.0.2
gunicorn==19.9.0
httplib2==0.12.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mysql-connector-python==8.0.15
nltk==3.4
numpy==1.16.2
oauth2==1.9.0.post1
protobuf==3.7.0
pyparsing==2.3.1
python-dateutil==2.8.0
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
singledispatch==3.4.0.3
six==1.12.0
sklearn==0.0
textblob==0.15.3
urllib3==1.24.1
Werkzeug==0.14.1
